requirements_test.txt
-r requirements.txt
build
GitPython
Nuitka
packaging
setuptools
twine
wheel
